Move unpack_ragged to analysis module
