+++ clouddrift/__init__.py
@@ -2,7 +2,7 @@ from ._version import version
 
 __version__ = version
 
-from clouddrift.raggedarray import RaggedArray, unpack_ragged
+from clouddrift.raggedarray import RaggedArray
 import clouddrift.adapters
 import clouddrift.analysis
 import clouddrift.haversine
