+++ clouddrift/raggedarray.py
@@ -401,52 +401,3 @@ class RaggedArray:
         """
         ak.to_parquet(self.to_awkward(), filename)
         return
-
-
-def unpack_ragged(
-    ragged_array: np.ndarray, rowsize: np.ndarray[int]
-) -> list[np.ndarray]:
-    """Unpack a ragged array into a list of regular arrays.
-
-    Unpacking a ``np.ndarray`` ragged array is about 2 orders of magnitude
-    faster than unpacking an ``xr.DataArray`` ragged array, so unless you need a
-    ``DataArray`` as the result, we recommend passing ``np.ndarray`` as input.
-
-    Parameters
-    ----------
-    ragged_array : array-like
-        A ragged_array to unpack
-    rowsize : array-like
-        An array of integers whose values is the size of each row in the ragged
-        array
-
-    Returns
-    -------
-    list
-        A list of array-likes with sizes that correspond to the values in
-        rowsize, and types that correspond to the type of ragged_array
-
-    Examples
-    --------
-
-    Unpacking longitude arrays from a ragged Xarray Dataset:
-
-    .. code-block:: python
-
-        lon = unpack_ragged(ds.lon, ds.rowsize) # return a list[xr.DataArray] (slower)
-        lon = unpack_ragged(ds.lon.values, ds.rowsize) # return a list[np.ndarray] (faster)
-
-    Looping over trajectories in a ragged Xarray Dataset to compute velocities
-    for each:
-
-    .. code-block:: python
-
-        for lon, lat, time in list(zip(
-            unpack_ragged(ds.lon.values, ds.rowsize),
-            unpack_ragged(ds.lat.values, ds.rowsize),
-            unpack_ragged(ds.time.values, ds.rowsize)
-        )):
-            u, v = velocity_from_position(lon, lat, time)
-    """
-    indices = np.insert(np.cumsum(np.array(rowsize)), 0, 0)
-    return [ragged_array[indices[n] : indices[n + 1]] for n in range(indices.size - 1)]
