+++ tests/analysis_tests.py
@@ -1,11 +1,12 @@
 from clouddrift.analysis import (
     apply_ragged,
     chunk,
-    regular_to_ragged,
+    position_from_velocity,
     ragged_to_regular,
+    regular_to_ragged,
     segment,
     subset,
-    position_from_velocity,
+    unpack_ragged,
     velocity_from_position,
 )
 from clouddrift.haversine import EARTH_RADIUS_METERS
@@ -19,6 +20,71 @@ from datetime import datetime, timedelta
 
 if __name__ == "__main__":
     unittest.main()
+
+
+def sample_ragged_array() -> RaggedArray:
+    drifter_id = [1, 2, 3]
+    rowsize = [5, 4, 2]
+    longitude = [[-121, -111, 51, 61, 71], [12, 22, 32, 42], [103, 113]]
+    latitude = [[-90, -45, 45, 90, 0], [10, 20, 30, 40], [10, 20]]
+    t = [[1, 2, 3, 4, 5], [2, 3, 4, 5], [4, 5]]
+    ids = [[1, 1, 1, 1, 1], [2, 2, 2, 2], [3, 3]]
+    test = [
+        [True, True, True, False, False],
+        [True, False, False, False],
+        [False, False],
+    ]
+    nb_obs = np.sum(rowsize)
+    nb_traj = len(drifter_id)
+    attrs_global = {
+        "title": "test trajectories",
+        "history": "version xyz",
+    }
+    variables_coords = ["ids", "time", "lon", "lat"]
+
+    coords = {"lon": longitude, "lat": latitude, "ids": ids, "time": t}
+    metadata = {"ID": drifter_id, "rowsize": rowsize}
+    data = {"test": test}
+
+    # append xr.Dataset to a list
+    list_ds = []
+    for i in range(0, len(rowsize)):
+        xr_coords = {}
+        for var in coords.keys():
+            xr_coords[var] = (
+                ["obs"],
+                coords[var][i],
+                {"long_name": f"variable {var}", "units": "-"},
+            )
+
+        xr_data = {}
+        for var in metadata.keys():
+            xr_data[var] = (
+                ["traj"],
+                [metadata[var][i]],
+                {"long_name": f"variable {var}", "units": "-"},
+            )
+
+        for var in data.keys():
+            xr_data[var] = (
+                ["obs"],
+                data[var][i],
+                {"long_name": f"variable {var}", "units": "-"},
+            )
+
+        list_ds.append(
+            xr.Dataset(coords=xr_coords, data_vars=xr_data, attrs=attrs_global)
+        )
+
+    ra = RaggedArray.from_files(
+        [0, 1, 2],
+        lambda i: list_ds[i],
+        variables_coords,
+        ["ID", "rowsize"],
+        ["test"],
+    )
+
+    return ra
 
 
 class chunk_tests(unittest.TestCase):
@@ -541,72 +607,7 @@ class apply_ragged_tests(unittest.TestCase):
 
 class subset_tests(unittest.TestCase):
     def setUp(self):
-        """
-        Create ragged array and output netCDF and Parquet file
-        """
-        drifter_id = [1, 2, 3]
-        rowsize = [5, 4, 2]
-        longitude = [[-121, -111, 51, 61, 71], [12, 22, 32, 42], [103, 113]]
-        latitude = [[-90, -45, 45, 90, 0], [10, 20, 30, 40], [10, 20]]
-        t = [[1, 2, 3, 4, 5], [2, 3, 4, 5], [4, 5]]
-        ids = [[1, 1, 1, 1, 1], [2, 2, 2, 2], [3, 3]]
-        test = [
-            [True, True, True, False, False],
-            [True, False, False, False],
-            [False, False],
-        ]
-        nb_obs = np.sum(rowsize)
-        nb_traj = len(drifter_id)
-        attrs_global = {
-            "title": "test trajectories",
-            "history": "version xyz",
-        }
-        variables_coords = ["ids", "time", "lon", "lat"]
-
-        coords = {"lon": longitude, "lat": latitude, "ids": ids, "time": t}
-        metadata = {"ID": drifter_id, "rowsize": rowsize}
-        data = {"test": test}
-
-        # append xr.Dataset to a list
-        list_ds = []
-        for i in range(0, len(rowsize)):
-            xr_coords = {}
-            for var in coords.keys():
-                xr_coords[var] = (
-                    ["obs"],
-                    coords[var][i],
-                    {"long_name": f"variable {var}", "units": "-"},
-                )
-
-            xr_data = {}
-            for var in metadata.keys():
-                xr_data[var] = (
-                    ["traj"],
-                    [metadata[var][i]],
-                    {"long_name": f"variable {var}", "units": "-"},
-                )
-
-            for var in data.keys():
-                xr_data[var] = (
-                    ["obs"],
-                    data[var][i],
-                    {"long_name": f"variable {var}", "units": "-"},
-                )
-
-            list_ds.append(
-                xr.Dataset(coords=xr_coords, data_vars=xr_data, attrs=attrs_global)
-            )
-
-        # create test ragged array
-        ra = RaggedArray.from_files(
-            [0, 1, 2],
-            lambda i: list_ds[i],
-            variables_coords,
-            ["ID", "rowsize"],
-            ["test"],
-        )
-
-        self.ds = ra.to_xarray()
+        self.ds = sample_ragged_array().to_xarray()
 
     def test_equal(self):
         ds_sub = subset(self.ds, {"test": True})
@@ -661,3 +662,23 @@ class subset_tests(unittest.TestCase):
 
         with self.assertRaises(ValueError):
             subset(self.ds, {"lon": (0, 180), "a": (0, 10)})
+
+
+class unpack_ragged_tests(unittest.TestCase):
+    def test_unpack_ragged(self):
+
+        ds = sample_ragged_array().to_xarray()
+
+        # Test unpacking into DataArrays
+        lon = unpack_ragged(ds.lon, ds.rowsize)
+
+        self.assertTrue(type(lon) is list)
+        self.assertTrue(np.all([type(a) is xr.DataArray for a in lon]))
+        self.assertTrue(np.all([lon[n].size == ds.rowsize[n] for n in range(len(lon))]))
+
+        # Test unpacking into np.ndarrays
+        lon = unpack_ragged(ds.lon.values, ds.rowsize)
+
+        self.assertTrue(type(lon) is list)
+        self.assertTrue(np.all([type(a) is np.ndarray for a in lon]))
+        self.assertTrue(np.all([lon[n].size == ds.rowsize[n] for n in range(len(lon))]))
